+++ core/txpool/ingress_filters_test.go
@@ -3,6 +3,7 @@ package txpool
 import (
 	"context"
 	"errors"
+	"math/big"
 	"net"
 	"testing"
 
@@ -14,6 +15,14 @@ import (
 )
 
 func TestInteropFilter(t *testing.T) {
+	// some placeholder transaction to test with
+	tx := types.NewTx(&types.DynamicFeeTx{
+		ChainID: big.NewInt(1),
+		Nonce:   1,
+		To:      &common.Address{},
+		Value:   big.NewInt(1),
+		Data:    []byte{},
+	})
 	t.Run("Tx has no logs", func(t *testing.T) {
 		logFn := func(tx *types.Transaction) ([]*types.Log, error) {
 			return []*types.Log{}, nil
@@ -24,7 +33,7 @@ func TestInteropFilter(t *testing.T) {
 		}
 		// when there are no logs to process, the transaction should be allowed
 		filter := NewInteropFilter(logFn, checkFn)
-		require.True(t, filter.FilterTx(context.Background(), &types.Transaction{}))
+		require.True(t, filter.FilterTx(context.Background(), tx))
 	})
 	t.Run("Tx errored when getting logs", func(t *testing.T) {
 		logFn := func(tx *types.Transaction) ([]*types.Log, error) {
@@ -34,16 +43,16 @@ func TestInteropFilter(t *testing.T) {
 			// make this return error, but it won't be called because logs retrieval errored
 			return errors.New("error")
 		}
-		// when log retrieval errors, the transaction should be allowed
+		// when log retrieval errors, the transaction should be denied
 		filter := NewInteropFilter(logFn, checkFn)
-		require.True(t, filter.FilterTx(context.Background(), &types.Transaction{}))
+		require.False(t, filter.FilterTx(context.Background(), tx))
 	})
 	t.Run("Tx has no executing messages", func(t *testing.T) {
 		logFn := func(tx *types.Transaction) ([]*types.Log, error) {
 			l1 := &types.Log{
 				Topics: []common.Hash{common.BytesToHash([]byte("topic1"))},
 			}
-			return []*types.Log{l1}, errors.New("error")
+			return []*types.Log{l1}, nil
 		}
 		checkFn := func(ctx context.Context, ems []interoptypes.Message, safety interoptypes.SafetyLevel) error {
 			// make this return error, but it won't be called because logs retrieval doesn't have executing messages
@@ -51,7 +60,7 @@ func TestInteropFilter(t *testing.T) {
 		}
 		// when no executing messages are included, the transaction should be allowed
 		filter := NewInteropFilter(logFn, checkFn)
-		require.True(t, filter.FilterTx(context.Background(), &types.Transaction{}))
+		require.True(t, filter.FilterTx(context.Background(), tx))
 	})
 	t.Run("Tx has valid executing message", func(t *testing.T) {
 		// build a basic executing message
@@ -80,7 +89,7 @@ func TestInteropFilter(t *testing.T) {
 		// when there is one executing message, the transaction should be allowed
 		// if the checkFn returns nil
 		filter := NewInteropFilter(logFn, checkFn)
-		require.True(t, filter.FilterTx(context.Background(), &types.Transaction{}))
+		require.True(t, filter.FilterTx(context.Background(), tx))
 		// confirm that one executing message was passed to the checkFn
 		require.Equal(t, 1, len(spyEMs))
 	})
@@ -111,7 +120,7 @@ func TestInteropFilter(t *testing.T) {
 		// when there is one executing message, and the checkFn returns an error,
 		// (ie the supervisor rejects the transaction) the transaction should be denied
 		filter := NewInteropFilter(logFn, checkFn)
-		require.False(t, filter.FilterTx(context.Background(), &types.Transaction{}))
+		require.False(t, filter.FilterTx(context.Background(), tx))
 		// confirm that one executing message was passed to the checkFn
 		require.Equal(t, 1, len(spyEMs))
 	})
@@ -174,129 +183,6 @@ func TestInteropFilterRPCFailures(t *testing.T) {
 			filter := NewInteropFilter(logFn, checkFn)
 			result := filter.FilterTx(context.Background(), &types.Transaction{})
 			require.Equal(t, false, result, "FilterTx result mismatch")
-		})
-	}
-}
-
-func TestInteropMessageFormatEdgeCases(t *testing.T) {
-	tests := []struct {
-		name          string
-		log           *types.Log
-		expectedError string
-	}{
-		{
-			name: "Empty Topics",
-			log: &types.Log{
-				Address: params.InteropCrossL2InboxAddress,
-				Topics:  []common.Hash{},
-				Data:    make([]byte, 32*5),
-			},
-			expectedError: "unexpected number of event topics: 0",
-		},
-		{
-			name: "Wrong Event Topic",
-			log: &types.Log{
-				Address: params.InteropCrossL2InboxAddress,
-				Topics: []common.Hash{
-					common.BytesToHash([]byte("wrong topic")),
-					common.BytesToHash([]byte("payloadHash")),
-				},
-				Data: make([]byte, 32*5),
-			},
-			expectedError: "unexpected event topic",
-		},
-		{
-			name: "Missing PayloadHash Topic",
-			log: &types.Log{
-				Address: params.InteropCrossL2InboxAddress,
-				Topics: []common.Hash{
-					common.BytesToHash(interoptypes.ExecutingMessageEventTopic[:]),
-				},
-				Data: make([]byte, 32*5),
-			},
-			expectedError: "unexpected number of event topics: 1",
-		},
-		{
-			name: "Too Many Topics",
-			log: &types.Log{
-				Address: params.InteropCrossL2InboxAddress,
-				Topics: []common.Hash{
-					common.BytesToHash(interoptypes.ExecutingMessageEventTopic[:]),
-					common.BytesToHash([]byte("payloadHash")),
-					common.BytesToHash([]byte("extra")),
-				},
-				Data: make([]byte, 32*5),
-			},
-			expectedError: "unexpected number of event topics: 3",
-		},
-		{
-			name: "Data Too Short",
-			log: &types.Log{
-				Address: params.InteropCrossL2InboxAddress,
-				Topics: []common.Hash{
-					common.BytesToHash(interoptypes.ExecutingMessageEventTopic[:]),
-					common.BytesToHash([]byte("payloadHash")),
-				},
-				Data: make([]byte, 32*4), // One word too short
-			},
-			expectedError: "unexpected identifier data length: 128",
-		},
-		{
-			name: "Data Too Long",
-			log: &types.Log{
-				Address: params.InteropCrossL2InboxAddress,
-				Topics: []common.Hash{
-					common.BytesToHash(interoptypes.ExecutingMessageEventTopic[:]),
-					common.BytesToHash([]byte("payloadHash")),
-				},
-				Data: make([]byte, 32*6), // One word too long
-			},
-			expectedError: "unexpected identifier data length: 192",
-		},
-		{
-			name: "Invalid Address Padding",
-			log: &types.Log{
-				Address: params.InteropCrossL2InboxAddress,
-				Topics: []common.Hash{
-					common.BytesToHash(interoptypes.ExecutingMessageEventTopic[:]),
-					common.BytesToHash([]byte("payloadHash")),
-				},
-				Data: func() []byte {
-					data := make([]byte, 32*5)
-					data[0] = 1 // Add non-zero byte in address padding
-					return data
-				}(),
-			},
-			expectedError: "invalid address padding",
-		},
-		{
-			name: "Invalid Block Number Padding",
-			log: &types.Log{
-				Address: params.InteropCrossL2InboxAddress,
-				Topics: []common.Hash{
-					common.BytesToHash(interoptypes.ExecutingMessageEventTopic[:]),
-					common.BytesToHash([]byte("payloadHash")),
-				},
-				Data: func() []byte {
-					data := make([]byte, 32*5)
-					data[32+23] = 1 // Add non-zero byte in block number padding
-					return data
-				}(),
-			},
-			expectedError: "invalid block number padding",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			var msg interoptypes.Message
-			err := msg.DecodeEvent(tt.log.Topics, tt.log.Data)
-			if tt.expectedError != "" {
-				require.Error(t, err)
-				require.ErrorContains(t, err, tt.expectedError)
-			} else {
-				require.NoError(t, err)
-			}
 		})
 	}
 }
